requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
docopt==0.6.2
idna==3.4
image==1.5.33
imgkit==1.2.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
pipreqs==0.4.11
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
yarg==0.1.9
